# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadow IT Detector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface CurrentUser {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  picture?: string | null;
  domain: string;
  isSubscribed: boolean;
  /** @nullable */
  trialEndsAt?: string | null;
}

export type OAuthAppRiskLevel = typeof OAuthAppRiskLevel[keyof typeof OAuthAppRiskLevel];


export const OAuthAppRiskLevel = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface OAuthApp {
  id: number;
  clientId: string;
  appName: string;
  category: string;
  riskLevel: OAuthAppRiskLevel;
  riskScore: number;
  userCount: number;
  scopes: string[];
  isDismissed: boolean;
  firstSeenAt: string;
  lastSeenAt: string;
  /** @nullable */
  iconUrl?: string | null;
}

export type OAuthAppDetail = OAuthApp & {
  authorizedUsers: string[];
  scopeDescriptions: string[];
  riskReasons: string[];
};

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Scan {
  id: number;
  status: ScanStatus;
  /** @nullable */
  appsFound?: number | null;
  /** @nullable */
  newAppsFound?: number | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface DashboardSummary {
  totalApps: number;
  highRiskApps: number;
  mediumRiskApps: number;
  lowRiskApps: number;
  newAppsThisWeek: number;
  dismissedApps: number;
  /** @nullable */
  lastScanAt?: string | null;
  totalUsers: number;
}

export interface CheckoutSession {
  url: string;
}

export interface BillingPortal {
  url: string;
}

export interface BillingStatus {
  isSubscribed: boolean;
  /** @nullable */
  plan?: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
  cancelAtPeriodEnd: boolean;
}

export type ListAppsParams = {
risk?: ListAppsRisk;
category?: string;
search?: string;
};

export type ListAppsRisk = typeof ListAppsRisk[keyof typeof ListAppsRisk];


export const ListAppsRisk = {
  high: 'high',
  medium: 'medium',
  low: 'low',
  all: 'all',
} as const;