# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadow IT Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalApps: number;
  highRiskApps: number;
  mediumRiskApps: number;
  lowRiskApps: number;
  newAppsThisWeek: number;
  dismissedApps: number;
  /** @nullable */
  lastScanAt?: string | null;
  totalUsers: number;
}