# lib/api-zod/src/generated/types/oAuthApp.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadow IT Detector API
 * OpenAPI spec version: 0.1.0
 */
import type { OAuthAppRiskLevel } from './oAuthAppRiskLevel';

export interface OAuthApp {
  id: number;
  clientId: string;
  appName: string;
  category: string;
  riskLevel: OAuthAppRiskLevel;
  riskScore: number;
  userCount: number;
  scopes: string[];
  isDismissed: boolean;
  firstSeenAt: string;
  lastSeenAt: string;
  /** @nullable */
  iconUrl?: string | null;
}